Remove unnecessary indirection to backend when communicating with native messenger
