+++ webextension/src/comms/native.ts
@@ -1,0 +1,133 @@
+import { browser } from 'webextension-polyfill-ts';
+import { EitherAsync } from 'purify-ts/EitherAsync';
+import { APP_NAME, MINIMUM_BINARY_VERSION } from 'Modules/config';
+import { compareAgainstMinimum } from 'Modules/semantic-versioning';
+
+interface CheckBinaryReq {
+	checkBinary: true;
+}
+
+interface GetBookmarksReq {
+	requestBookmarks: true;
+}
+
+interface SaveBookmarkReq {
+	saveBookmark: true;
+	bookmark: LocalBookmarkUnsaved;
+}
+
+interface UpdateBookmarkReq {
+	updateBookmark: true;
+	bookmark: LocalBookmark;
+}
+
+interface DeleteBookmarkReq {
+	deleteBookmark: true;
+	bookmarkId: RemoteBookmark['id'];
+}
+
+export type NativeRequest =
+	CheckBinaryReq | GetBookmarksReq | SaveBookmarkReq | UpdateBookmarkReq | DeleteBookmarkReq;
+
+type CheckBinaryRes =
+	| { outdatedBinary: true }
+	| { cannotFindBinary: true }
+	| { unknownError: true };
+
+interface GetBookmarksRes {
+	bookmarksUpdated: true;
+}
+
+interface SaveBookmarkRes {
+	bookmarkSaved: true;
+}
+
+interface UpdateBookmarkRes {
+	bookmarkUpdated: true;
+}
+
+interface DeleteBookmarkRes {
+	bookmarkDeleted: true;
+}
+
+export type NativeResponse =
+	CheckBinaryRes | GetBookmarksRes | SaveBookmarkRes | UpdateBookmarkRes | DeleteBookmarkRes;
+
+export enum NativeRequestMethod {
+	GET = 'GET',
+	OPTIONS = 'OPTIONS',
+	POST = 'POST',
+	PUT = 'PUT',
+	DELETE = 'DELETE',
+}
+
+interface ErrResponse {
+	success: false;
+	message: string;
+}
+
+type NativeGETResponse = {
+	success: true;
+	bookmarks: RemoteBookmark[];
+} | ErrResponse;
+
+interface NativeOPTIONSResponse {
+	success: true;
+	binaryVersion: string;
+}
+
+type NativePOSTResponse = { success: true; id: number } | { success: false };
+
+interface NativePUTResponse {
+	success: boolean;
+}
+
+interface NativeDELETEResponse {
+	success: boolean;
+}
+
+interface NativeRequestData {
+	GET: undefined;
+	OPTIONS: undefined;
+	POST: { bookmark: RemoteBookmarkUnsaved };
+	PUT: { bookmark: RemoteBookmark };
+	DELETE: { bookmark_id: RemoteBookmark['id'] };
+}
+
+interface NativeRequestResult {
+	GET: NativeGETResponse;
+	OPTIONS: NativeOPTIONSResponse;
+	POST: NativePOSTResponse;
+	PUT: NativePUTResponse;
+	DELETE: NativeDELETEResponse;
+}
+
+const sendMessageToNative = <T extends NativeRequestMethod>(method: T, data: NativeRequestData[T]) =>
+	browser.runtime.sendNativeMessage(APP_NAME, { method, data }) as Promise<NativeRequestResult[T]>;
+
+// Ensure binary version is equal to or newer than what we're expecting, but on
+// the same major version (semantic versioning)
+export const checkBinaryVersionFromNative = () => EitherAsync<Error, void>(() => sendMessageToNative(NativeRequestMethod.OPTIONS, undefined)
+	.then((res) => {
+		const valid = !!(
+			res &&
+			res.success &&
+			res.binaryVersion &&
+			compareAgainstMinimum(MINIMUM_BINARY_VERSION, res.binaryVersion)
+		);
+
+		if (!valid) throw new Error('Incompatible version');
+	}));
+
+export const getBookmarksFromNative = () =>
+	sendMessageToNative(NativeRequestMethod.GET, undefined);
+
+export const saveBookmarkToNative = (bookmark: RemoteBookmarkUnsaved) =>
+	sendMessageToNative(NativeRequestMethod.POST, { bookmark });
+
+export const updateBookmarkToNative = (bookmark: RemoteBookmark) =>
+	sendMessageToNative(NativeRequestMethod.PUT, { bookmark });
+
+export const deleteBookmarkFromNative = (bookmarkId: RemoteBookmark['id']) =>
+	// eslint-disable-next-line @typescript-eslint/camelcase
+	sendMessageToNative(NativeRequestMethod.DELETE, { bookmark_id: bookmarkId });
