+++ webextension/src/pages/content/container.tsx
@@ -1,9 +1,8 @@
 import React from 'react';
 import { connect } from 'react-redux';
-import { requestBookmarks } from 'Comms/frontend';
 import { AppState } from 'Store';
 import { setLimitNumRendered, setAddBookmarkModalDisplay } from 'Store/bookmarks/actions';
-import { openAllFilteredBookmarksAndExit } from 'Store/epics';
+import { openAllFilteredBookmarksAndExit, syncBookmarks } from 'Store/epics';
 import { getNumFilteredUnrenderedBookmarks } from 'Store/selectors';
 import ContentPage from './content';
 
@@ -17,7 +16,7 @@ const ContentContainer: Comp<Props> = props => (
 		onEnableLimitlessRender={props.setLimitNumBookmarksRendered.bind(null, false)}
 		toggleAddBookmarkForm={props.setAddBookmarkModalDisplay.bind(null, !props.displayAddBookmarkModal)}
 		openAllFilteredBookmarksWithoutConfirmation={props.openAllFilteredBookmarks}
-		refreshBookmarks={requestBookmarks}
+		refreshBookmarks={props.syncBookmarks}
 		numRemainingBookmarks={props.numRemainingBookmarks}
 		displayTutorialMessage={props.displayTutorialMessage}
 	/>
@@ -30,6 +29,7 @@ const mapStateToProps = (state: AppState) => ({
 });
 
 const mapDispatchToProps = {
+	syncBookmarks,
 	setAddBookmarkModalDisplay,
 	openAllFilteredBookmarks: openAllFilteredBookmarksAndExit,
 	setLimitNumBookmarksRendered: setLimitNumRendered,
